+++ src/workspace.rs
@@ -41,45 +41,32 @@ pub fn did_change_configuration(meta: EditorMeta, mut params: EditorParams, ctx:
 
 pub fn configuration(
     meta: EditorMeta,
-    params: Params,
+    _params: Params,
     ctx: &mut Context,
 ) -> Result<Value, jsonrpc_core::Error> {
-    let params = params.parse::<ConfigurationParams>()?;
+    // We ignore our parameters because they are not useful.
+    // The "section" parameter is used to identify the language server, but we already know
+    // that from the filetype. Since we simply send initialization options, we actually don't
+    // want users to put section names in there, because it would make the initialization
+    // options unusable for initialization and for didChangeConfiguration.
+    // The "scopeUri" parameter gives a file/directory that the config should apply to,
+    // but kak-lsp doesn't have a concept of per-file configuration, and workspaces
+    // are separated by kak-lsp processes.
 
     let dynamic_config = request_dynamic_configuration_from_kakoune(&meta, ctx);
     let settings = match dynamic_config
-        .as_ref()
-        .and_then(|cfg| cfg.initialization_options.as_ref())
+        .and_then(|cfg| cfg.initialization_options)
         .or_else(|| {
             ctx.config
                 .language
                 .get(&ctx.language_id)
-                .and_then(|conf| conf.initialization_options.as_ref())
+                .and_then(|conf| conf.initialization_options.clone())
         }) {
         Some(settings) => settings,
         None => return Ok(Value::Array(Vec::new())),
     };
 
-    let items = params
-        .items
-        .iter()
-        .map(|item| {
-            // There's also a `scopeUri`, which lists the file/folder
-            // that the config should apply to. But kak-lsp doesn't
-            // have a concept of per-file configuration and workspaces
-            // are separated by kak-lsp processes.
-            item.section
-                .as_ref()
-                // The specification isn't clear about whether you should
-                // reply with just the value or with `json!({ section: <value> })`.
-                // Tests indicate the former.
-                .and_then(|section| settings.get(section))
-                .cloned()
-                .unwrap_or(Value::Null)
-        })
-        .collect::<Vec<Value>>();
-
-    Ok(Value::Array(items))
+    Ok(Value::Array(vec![settings]))
 }
 
 pub fn workspace_symbol(meta: EditorMeta, params: EditorParams, ctx: &mut Context) {