+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Bug fixes:
 - Fix renaming of Rust lifetimes (#474).
 - The suggested config for `rust-analyzer` was fixed for the case that `rustup` is installed but `rust-analyzer` is not installed via `rustup`.
 - Fix spurious cursor movement on `lsp-rename` and `lsp-rename-prompt` (#504).
+- When responding to `workspace/configuration`, kak-lsp has already used the configured initialization options, but now ignores the `section` parameter. Consequently, the section should not be included in the initialization options, which makes the options usable for other requests (#507).
 
 Deprecations:
 - `%opt{lsp_server_initialization_options}` and `%opt{lsp_server_configuration}` are deprecated in favor of setting `[language.<filetype>.initialization_options]` in `%opt{lsp_config}`(#500).